fixed bean-injection for RestTemplate; added codecov;
