+++ README.md
@@ -1,2 +1,3 @@
 [![License: GPL v3](https://img.shields.io/badge/License-GPL%20v3-blue.svg)](http://www.gnu.org/licenses/gpl-3.0)
 [![Build Status](https://travis-ci.org/nt-ca-aqe/marketing-metrics-collector.svg?branch=master)](https://travis-ci.org/nt-ca-aqe/marketing-metrics-collector)
+[![codecov](https://codecov.io/gh/nt-ca-aqe/marketing-metrics-collector/branch/master/graph/badge.svg)](https://codecov.io/gh/nt-ca-aqe/marketing-metrics-collector)
