+++ mmc-commons/src/main/java/info/novatec/metricscollector/commons/CommonsApplicationInitializer.java
@@ -3,26 +3,17 @@ package info.novatec.metricscollector.commons;
 import org.influxdb.InfluxDBFactory;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.context.annotation.Bean;
-import org.springframework.web.client.RestTemplate;
+
+import lombok.Setter;
 
 
 @SpringBootApplication
 @ConfigurationProperties(prefix = "influx")
 public class CommonsApplicationInitializer {
 
-    @lombok.Setter
+    @Setter
     String url;
-
-//    public static void main(String[] args) {
-//        SpringApplication.run(ApplicationInitializerCommons.class, args);
-//    }
-
-    @Bean
-    public RestTemplate restTemplate() {
-        return new RestTemplateBuilder().build();
-    }
 
     @Bean
     public InfluxService influxService() {
