+++ mmc-github/src/main/java/info/novatec/metricscollector/github/GithubApplicationInitializer.java
@@ -6,9 +6,11 @@ import java.util.List;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Import;
 import org.springframework.scheduling.annotation.EnableScheduling;
+import org.springframework.web.client.RestTemplate;
 
 import lombok.Getter;
 import lombok.Setter;
@@ -35,6 +37,11 @@ public class GithubApplicationInitializer {
     @Bean
     public String token(){
         return token;
+    }
+
+    @Bean
+    public RestTemplate restTemplate() {
+        return new RestTemplateBuilder().build();
     }
 
 }
