+++ mmc-github/src/test/unit/java/info/novatec/metricscollector/github/GithubApplicationInitializerTest.java
@@ -7,6 +7,7 @@ import javax.annotation.PostConstruct;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.test.context.junit4.SpringRunner;
+import org.springframework.web.client.RestTemplate;
 
 
 @RunWith(SpringRunner.class)
@@ -28,4 +29,9 @@ public class GithubApplicationInitializerTest {
         assertThat(token).isEqualTo(TOKEN);
     }
 
+    @Test
+    public void createRestTemplate(){
+        RestTemplate influxService = initializer.restTemplate();
+        assertThat(influxService).isNotNull();
+    }
 }
