+++ mmc-google-analytics/src/main/resources/application.yml-template
@@ -1,0 +1,9 @@
+influx:
+    url:
+    dbName:
+    retention:
+analytics:
+    applicationName:
+    viewId:
+    keyFileLocation:
+    hostName:
